Adding binomial namespace, rename function and filename, correcting comments
